Refactor board methods into separate modules for better organization
