+++ V2/js/board/effects.js
@@ -1,0 +1,14 @@
+export function triggerScreenShake(magnitude, duration) {
+    this.screenShake = {
+        startTime: performance.now(),
+        magnitude: magnitude,
+        duration: duration
+    };
+}
+
+export function triggerAttackEffect() {
+    this.attackEffect = {
+        startTime: performance.now(),
+        duration: 400
+    };
+}
